Backend/legal_researcher/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.10.0
pydantic-settings>=2.3.0
python-multipart==0.0.9
PyPDF2==3.0.1
firecrawl-py>=1.0.0
langchain-groq
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
requests==2.32.3
gunicorn==22.0.0
langchain-community
langchain
langgraph==0.2.67
google-auth-oauthlib
google-api-python-client
fpdf2==2.8.2
deep-translator==1.11.4
transformers
datasets
evaluate
sacrebleu
sentencepiece
torch
torchtext
libsql-experimental==0.0.55
libsql-client
